feat: Presto grammar from antlr4 (#31)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/verax/pull/31

Check in the generated ANTLR4 parser, lexer, etc based on Presto Java's grammar. The parser lives under `axiom/sql/presto` with namespace `facebook::velox::sql`.

For now, I am checking in the auto-generated code from ANTLR4.

The end usage of this checked-in code is for users to implement the visitors so that AST structures can be created to logically represent the ANTLR parse tree.

Follow up PRs:
- Use ANTLR4 function in CMAKE to automatically generate the parsed files. See https://github.com/antlr/antlr4/tree/dev/runtime/Cpp/cmake

- Implement some of the callbacks to create the AST

- Basic/primitive planner/binder which converts the AST into Masha's LogicalPlanNode

Reviewed By: oerling, feilong-liu

Differential Revision: D78024793

fbshipit-source-id: 4108d84ab2e87ab2f165012ce8a3fbc78ac0ea5e