feat: Add a multidimensional interpolation model (#104)
Summary:
Adds a multidimensional cube for representing a a scalar field in a n-dimensional Euclidian space. Each point with a non-negative coordinate in each dimension has a float value derived from neighboring points.

This is used for a cost model configured by a set of measurements. The dimensions are typically things like number of rows, number of distinct values, number of columns etc. The measure is typically a throughput or a time per value or a size.

Given a set of points we fill in a cube. Suppose a two dimensional cube with measures alon both axis, as in [1, 10], [1, 20], [1, 40] for the y axis and [1, 1], [20, 1], [100, 1] for the x axis. We will in values for all combinations of coordinates for which there is no measure. For example [[10, 20] gets a value based on the slope along the x axis from 1m to 20 and from 1 to 10 along the y axis.

The cube is filled at setup  time. At query time we use linear interpolation from closest neighbors along each axis to estimate the measure at the coordinate.

We normalize the dimension along the lowest and highest value of each axis to be in the range [0, 1]. We likewise normalize the value of the point we are querying. If this falls inside the cube along each dimension, the point has 2^rank neighbors for which a measure is known or precomputed. We weigh the importance of each neighbor by 1/distance from the neighbor to the point being queried. For a precise match along each dimension we return the measure at the point. For other cases we return the sum of measure * weight / sum of weights.

For a point outside of the cube, we take the slope at the edge of the cube along each dimension that is outside the cube. For the dimensions that are inside the cube we interpolate the value at the face of the cube based on the coordinates that are within [0, 1]. We continue the slope from that point along each dimension that is not within [0, 1].

The dimensions are real world quantities such as counts and are non-negative. We place the origin of the cube at the lowest value along each dimension for which there is a measure.  We allow queries for points with a non-negative coordinate for the normalized dimension, i.e. where the value is >= to the lowest value of the coordinate in the initial set of measures.

The functions modeled are expected to have a non-negative derivative along each dimension. The slope may vary. Cliffs like moving from one algorithm to another can be represented by a steep slope between nearby points but the model always has a slope less than infinity. Only the nearest values along each dimension participate in the interpolation, so a nearby cliff does not bias the estimate. unless the coordinate being queried is actually between the end points of the cliff.

mdm.add

git add axiom/optimizer/Model.cpp
git add axiom/optimizer/Model.h
git add axiom/optimizer/ModelTest.cpp

Pull Request resolved: https://github.com/facebookexperimental/verax/pull/104

Reviewed By: Yuhta

Differential Revision: D79536111

Pulled By: oerling

fbshipit-source-id: 4bdf571c4317e073395d18d46c2e44dfed1e2cba